refactor code to remove circular imports
