+++ backend/Organisation/OrganisationHelpers.py
@@ -1,11 +1,13 @@
 from typing import List
 
+from backend.Organisation.TeamHelpers import getTeamsById
 from backend.States.Enums import Roles
 from backend.States.Errors import AuthenticationError
 from backend.States.IOSchema import Organisation, Meeting, Person, InviteOutput
 from backend.database.database import mapOrgNameToID, mapOrgIDToName, mapTeamNameToId, getUserOrgs, getBulkUsersByIds, \
     getOrgRoleByID, getTeamRoleByID, getAll, \
-    getUserDetailsByEmail, getInvites, getInvitesByUser, addUserToOrg, getStatusOrg, getStatusTeam, inOrg, removeUser
+    getUserDetailsByEmail, getInvites, getInvitesByUser, addUserToOrg, getStatusOrg, getStatusTeam, inOrg, removeUser, \
+    removeTeamUser
 
 
 def getOrganisationsByID(orgIds: [int] = None) -> [Organisation]:
@@ -139,3 +141,16 @@ def forceJoin(email: str):
 def isUserInOrg(userId: int, orgId: int) -> bool:
     return inOrg(userId, orgId) is not None
 
+
+def removeUserUnchecked(userId, org):
+    """
+    Private method,
+    only to be called internally,
+    does not do error checking
+    :param userId:
+    :param org:
+    """
+    teams = getTeamsById(org, userId)
+    for team in teams:
+        removeTeamUser(userId, org, team.id)
+    removeUser(userId, org)
