+++ backend/Organisation/Organisations.py
@@ -2,19 +2,18 @@ from typing import List
 
 from fastapi import HTTPException
 
-from backend.States.Enums import Roles
-from backend.States.IOSchema import Organisation, OrganisationReport, OrganisationPersonalReport, Person, Team, \
-    TeamPersonalReport, \
-    TeamReport, OrgTeam, Meeting, InviteOutput, MeetingInput
-from backend.Organisation.OrganisationHelpers import getOrganisationsByID, getOrganisationByName, getOrganisationByID, \
-    getTeamByName, \
-    meetify, getRoleByID, getTRoleByID, getAllUsers, getPendingInvites, getStatus, getTeamStatus, isUserInOrg
+from backend.Organisation.OrganisationHelpers import getOrganisationByName, getTeamByName, \
+    meetify, getRoleByID, getTRoleByID, getPendingInvites, getStatus, isUserInOrg, removeUserUnchecked
+from backend.Organisation.Team import getTeamsById
 from backend.Profile.UserAccounts import getUserByID
+from backend.States.Enums import Roles
 from backend.States.Errors import AuthenticationError
-from backend.database.database import mapOrgIDToName, mapOrgNameToID, getUserOrgs, getTeamsByOrg, getMeetingsByTeam, \
+from backend.States.IOSchema import Organisation, OrganisationReport, OrganisationPersonalReport, Person, OrgTeam, \
+    Meeting, InviteOutput
+from backend.database.database import getMeetingsByTeam, \
     getMeetingsByOrg, \
-    makeTeam, teamExists, addUserToTeam, existsOrganisation, makeOrganisation, getOwner, addUserToOrg, \
-    getTeamsByOrgStatus, setActiveOrganisation, deleteOrganisationByID, removeUser, removeTeamUser
+    addUserToTeam, existsOrganisation, makeOrganisation, getOwner, addUserToOrg, \
+    setActiveOrganisation, deleteOrganisationByID
 
 
 def createOrganisation(OrganisationName: str, OwnerID: int) -> Organisation:
@@ -54,24 +53,6 @@ def getOrganisationReport(UserID: int, OrganisationName: str) -> OrganisationPer
     return report
 
 
-def getTeamReport(userID: int, teamName: str, organisationName: str) -> TeamPersonalReport:
-    organisation = getOrganisationByName(organisationName)
-    team: int = getTeamByName(orgId=organisation, teamName=teamName)
-    userRole = getTRoleByID(organisation, team, userID)
-    if not userRole:
-        AuthenticationError("User is not in team")
-    admins: [Person] = getTeamStatus(organisation, team, Roles.ADMIN)
-    users: [Person] = getTeamStatus(organisation, team, Roles.USER)
-    allUsers: [Person] = getAllUsers(organisation, team)
-    otherUsers = filter(lambda x: (x not in users) and (x not in admins), allUsers)
-    teamReport = TeamReport(id=team, name=teamName,
-                            admins=admins,
-                            users=users,
-                            otherUsers=otherUsers)
-    report = TeamPersonalReport(isPermitted=True, userRole=userRole, team=teamReport)
-    return report
-
-
 def getMeetings(userId: int, orgTeam: OrgTeam):
     org: int = getOrganisationByName(orgTeam.organisation)
     team: int = getTeamByName(org, orgTeam.name)
@@ -89,28 +70,6 @@ def getAllMeetings(userId: int, orgName: str) -> [Meeting]:
     return meetify(meetings)
 
 
-def getTeamsById(id: int, UserId: int, status: Roles | None = None):
-    if not status:
-        teams = getTeamsByOrg(id, UserId)
-    else:
-        teams = getTeamsByOrgStatus(id, UserId, status)
-    teammer = lambda row: Team(id=row[0], name=row[1])
-    teams = list(map(teammer, teams))
-    return teams
-
-
-def getTeams(name: str, Userid: int, status: Roles | None = None):
-    id = getOrganisationByName(name)
-    if not isUserInOrg(Userid, id):
-        AuthenticationError("You can only see your own organisation's teams")
-    return getTeamsById(id, Userid, status)
-
-
-'''
-⚠️ if TeamControl is enabled and existing user is added they lose their powers
-'''
-
-
 def addUser(id: int, organisation: str, userId: int, role: str, teamName: str = None) -> Person:
     organisation = getOrganisationByName(organisation)
     teamName = getTeamByName(organisation, teamName)
@@ -120,34 +79,6 @@ def addUser(id: int, organisation: str, userId: int, role: str, teamName: str = 
     addUserToTeam(organisation, userId, role, teamName, role)
     user = getUserByID(userId).user
     return Person(id=userId, username=user.username, email=user.email, firstName=user.firstName, lastName=user.lastName)
-
-
-def createTeam(userId: int, orgteam: OrgTeam):
-    org = getOrganisationByName(orgteam.organisation)
-    if teamExists(org, orgteam.name) is not None:
-        raise HTTPException(status_code=400, detail="Team already exists")
-    if getRoleByID(org, userId) == Roles.USER.value:
-        raise HTTPException(status_code=403, detail="User is not authorized to create team")
-    makeTeam(org, orgteam.name)
-    id = getTeamByName(org, orgteam.name)
-    owner = getOwner(org)[0]
-    addUserToTeam(org, owner, Roles.ADMIN.value, id, Roles.ADMIN.value)
-    addUserToTeam(org, userId, Roles.ADMIN.value, id, Roles.ADMIN.value)
-    return id
-
-
-def removeUserUnchecked(userId, org):
-    """
-    Private method,
-    only to be called internally,
-    does not do error checking
-    :param userId:
-    :param org:
-    """
-    teams = getTeamsById(org, userId)
-    for team in teams:
-        removeTeamUser(userId, org, team.id)
-    removeUser(userId, org)
 
 
 def removeUserOrg(userId: int, org: str, remover: int):
@@ -176,35 +107,3 @@ def removeUserOrg(userId: int, org: str, remover: int):
     else:
         print(f"User {userId} has been removed from the organisation {org} by User {remover}")
     removeUserUnchecked(userId, org)
-
-
-def removeUserTeam(userId: int, orgteam: OrgTeam, remover: int):
-    """
-    Removes user from the organisation
-    if user is  remover, user just quits the organisation
-    :param orgteam: pair of team and its org to remove user from
-    :param userId: user to be removed
-    :param remover: admin removing user
-    """
-    org: int = getOrganisationByName(orgteam.organisation)
-    if not org:
-        raise HTTPException(status_code=404, detail="Organisation not found")
-
-    role = (getRoleByID(org, remover))
-    if role is None:
-        raise HTTPException(status_code=404, detail=f"User {remover} not in organisation")
-    if remover != userId and role == Roles.USER:
-        raise HTTPException(status_code=403, detail="User is not authorised to remove user")
-    if not isUserInOrg(userId, org):
-        raise HTTPException(status_code=404, detail=f"User {userId} not in organisation")
-
-    team: int = getTeamByName(org, orgteam.name)
-    if not team:
-        raise HTTPException(status_code=404, detail="Team not found")
-    #TODO:check for team membership
-    if userId == remover:
-        #TODO: do proper logging
-        print(f"User {userId} has left the team {orgteam.name} in organisation {org}")
-    else:
-        print(f"User {userId} has been removed from the team {orgteam.name} in organisation {org} by User {remover}")
-    removeTeamUser(userId, org, team)
